pitch-visualizer/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
jinja2==3.1.3
python-multipart==0.0.9
python-dotenv==1.0.1

# NLP
nltk==3.8.1

# LLM Prompt Engineering
google-generativeai==0.5.4

# HTTP (image generation)
requests==2.31.0
